New getdss version from Bill Gary, make it work on Linux x64 .

+++ trunk/skychart/library/getdss/hdcmprss.cpp
@@ -12,7 +12,7 @@ int decode( int filesize, char *file_buff,
 #define ZKEY_CLOCK  0
 #endif
 
-extern long times[];
+extern int32_t times[];
 
         /* (*a)[nx][ny] is the image array  */
         /* Note that ny is the fast-varying dimension   */
@@ -22,14 +22,15 @@ extern int hdecompress( int **a, int *nx, int *ny, int filesize,
                             char *file_buff)
 {
    int scale, rval;
-   long t0;
+   int32_t t0;
 
    t0 = ZKEY_CLOCK;
    dss_debug_printf( "decode, ");
    rval = decode( filesize, file_buff, a, nx, ny, &scale);
-   //times[1] += ZKEY_CLOCK - t0;
+//   times[1] += ZKEY_CLOCK - t0;
    if( rval)
       return( rval);
+
    t0 = ZKEY_CLOCK;
    dss_debug_printf( "scale, ");
    if( scale > 1)
@@ -39,12 +40,12 @@ extern int hdecompress( int **a, int *nx, int *ny, int filesize,
       while( p < endptr)
          *p++ *= scale;
       }
-   //times[2] += ZKEY_CLOCK - t0;
+//   times[2] += ZKEY_CLOCK - t0;
 
    dss_debug_printf( "hinv, ");
    t0 = ZKEY_CLOCK;
    rval = hinv( *a,*nx,*ny);           /* Inverse H-transform  */
-   //times[3] += ZKEY_CLOCK - t0;
+//   times[3] += ZKEY_CLOCK - t0;
    dss_debug_printf( "done. ");
    return( rval);
 }
